Fix WinWing PFP3N line problem and multiple CDUs
