--- MobiFlight/Joysticks/Winwing/Core/MsgEntry.cs
@@ -1,9 +1,0 @@
-﻿namespace MobiFlight.Joysticks.Winwing
-{
-    internal class MsgEntry
-    {
-        public byte StartPos;
-        public byte[] Mask;
-        public byte[] Data;
-    }
-}
