+++ MobiFlight/Joysticks/Winwing/WinwingCdu.cs
@@ -1,4 +1,5 @@
 ﻿using MobiFlight.Config;
+using MobiFlightWwFcu;
 using System;
 using System.Collections.Generic;
 using WebSocketSharp.Server;
